# App/schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sumradio",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/runs": {
      "get": {
        "summary": "Runs",
        "operationId": "runs_api_runs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create",
        "operationId": "create_api_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/runs/{run_id}/locations": {
      "get": {
        "summary": "Run Locations",
        "operationId": "run_locations_api_runs__run_id__locations_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Location-Output"
                  },
                  "title": "Response Run Locations Api Runs  Run Id  Locations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "summary": "Get Run",
        "operationId": "get_run_api_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/runs/{run_id}/recording/start": {
      "post": {
        "summary": "Start Recording",
        "operationId": "start_recording_api_runs__run_id__recording_start_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordingStart"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/recording/stop": {
      "post": {
        "summary": "Stop Recording",
        "operationId": "stop_recording_api_runs__run_id__recording_stop_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/communications/{comm_id}/corrections": {
      "post": {
        "summary": "Correct",
        "operationId": "correct_api_runs__run_id__communications__comm_id__corrections_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "comm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comm Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/communications/{comm_id}/retry": {
      "post": {
        "summary": "Retry",
        "operationId": "retry_api_runs__run_id__communications__comm_id__retry_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "comm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comm Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Versioned"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/tasks": {
      "post": {
        "summary": "Create Task",
        "operationId": "create_task_api_runs__run_id__tasks_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/runs/{run_id}/tasks/{task_id}": {
      "put": {
        "summary": "Edit Task",
        "operationId": "edit_task_api_runs__run_id__tasks__task_id__put",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskEdit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/tasks/{task_id}/transition": {
      "post": {
        "summary": "Transition",
        "operationId": "transition_api_runs__run_id__tasks__task_id__transition_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/tasks/{task_id}/location": {
      "post": {
        "summary": "Location",
        "operationId": "location_api_runs__run_id__tasks__task_id__location_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmLocation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/tasks/{task_id}/interpretation-location": {
      "post": {
        "summary": "Interpretation Location",
        "operationId": "interpretation_location_api_runs__run_id__tasks__task_id__interpretation_location_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmInterpretationLocation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/audio/{comm_id}/{variant}": {
      "get": {
        "summary": "Audio",
        "operationId": "audio_api_runs__run_id__audio__comm_id___variant__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "comm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comm Id"
            }
          },
          {
            "name": "variant",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Variant"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CorrectRequest": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "actor": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Actor"
          },
          "text": {
            "type": "string",
            "maxLength": 100000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version",
          "actor",
          "text"
        ],
        "title": "CorrectRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Location-Output": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "municipality": {
            "type": "string",
            "title": "Municipality"
          },
          "aliases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Aliases"
          },
          "lat": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lon"
          },
          "precision": {
            "type": "string",
            "enum": [
              "representative",
              "specific"
            ],
            "title": "Precision",
            "default": "representative"
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "default": ""
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url",
            "default": ""
          },
          "checked_at": {
            "type": "string",
            "title": "Checked At",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "municipality",
          "aliases",
          "lat",
          "lon",
          "precision",
          "detail",
          "source",
          "source_url",
          "checked_at"
        ],
        "title": "Location"
      },
...
      "RecordingStart": {
        "properties": {
          "device": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device"
          },
          "manual_mode": {
            "type": "boolean",
            "title": "Manual Mode",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RecordingStart"
      },
      "Region-Input": {
        "properties": {
          "prefecture": {
            "type": "string",
            "maxLength": 10,
            "minLength": 3,
            "title": "Prefecture"
          },
          "municipality": {
            "type": "string",
            "maxLength": 50,
            "title": "Municipality",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "prefecture"
        ],
        "title": "Region"
      },
...
      "Run": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "demo_profile_id": {
            "type": "string",
            "title": "Demo Profile Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "region": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Region-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "default": 1
          },
          "communications": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Communication"
            },
            "type": "object",
            "title": "Communications"
          },
          "tasks": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Task"
            },
            "type": "object",
            "title": "Tasks"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/History"
            },
            "type": "array",
            "title": "History"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "demo_profile_id",
          "title",
          "region",
          "created_at",
          "version",
          "communications",
          "tasks",
          "history"
        ],
        "title": "Run"
      },
      "RunCreate": {
        "properties": {
          "demo_profile_id": {
            "type": "string",
            "title": "Demo Profile Id"
          },
          "title": {
            "type": "string",
            "maxLength": 100,
            "title": "Title",
            "default": ""
          },
          "region": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Region-Input"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "demo_profile_id"
        ],
        "title": "RunCreate"
      },
      "RunView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "demo_profile_id": {
            "type": "string",
            "title": "Demo Profile Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "region": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Region-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "default": 1
          },
          "communications": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Communication"
            },
            "type": "object",
            "title": "Communications"
          },
          "tasks": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Task"
            },
            "type": "object",
            "title": "Tasks"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/History"
            },
            "type": "array",
            "title": "History"
          },
          "interpretation_locations": {
            "items": {
              "$ref": "#/components/schemas/InterpretationLocationOption"
            },
            "type": "array",
            "title": "Interpretation Locations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "demo_profile_id",
          "title",
          "region",
          "created_at",
          "version",
          "communications",
          "tasks",
          "history",
          "interpretation_locations"
        ],
        "title": "RunView"
      },
...
      "TransitionRequest": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "actor": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Actor"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version",
          "actor",
          "status"
        ],
        "title": "TransitionRequest"
      },
...
      "Versioned": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "actor": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Actor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version",
          "actor"
        ],
        "title": "Versioned"
      }
    }
  }
}